requirement.txt
streamlit
google-genai>=1.51.0
Pillow
pypdf2